Package bot commands
