+++ .gitignore
@@ -15,3 +15,4 @@ secret.auto.tfvars
 
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
+main
